+++ objC/kpRectangle/kpRectangle/kpSquare.m
@@ -10,12 +10,12 @@
 
 @implementation kpSquare: kpRectangle
    
--(void) setSide: (int) s{
+-(void) setSide: (float) s{
     
     [self setWidth: s andHeight: s];
 }
 
--(int) side{
+-(float) side{
     
     return self.width;
 }
